optimation
if resp_status == http_client.FORBIDDEN:
        # If there's no details about the 403 type, don't retry.
        if not content:
            return False
The code block looks correct, but I noticed that the code block inside the if not content: line is not complete. This block of code needs an operation in progress or returning a value. Perhaps you could add an expression that returns True or False.

        else:
            reason = json.loads(content).get('error', {}).get('errors', [{}])[0].get('reason', None)
            if reason in _RETRY_REASONS:
                return True
    return False